+++ server/controllers/decks.controller.js
@@ -22,16 +22,17 @@ class DecksController {
             // NEW USER WITH BODY PROPERTIES
             let newDeck = new Deck({
                 name: req.body.name,
-                user: req.user._id, // TODO: implement passport
+                user: req.user._id,
                 cards: []
             });
-            
+
             for (let i = 0; i < cards.length; i++) {
                 newDeck.cards.push(cards[i]._id)
             }
-            
+
             await newDeck.save();
-            res.status(203).send({ success: `Varalho ${req.body.name} registada` })
+            const deck = await Deck.findById(newDeck._id).populate("cards") // Better way? 
+            res.status(203).send({ success: `Baralho ${req.body.name} registado`, deck })
 
         } catch (error) {
             res.status(400).send(error)
@@ -40,9 +41,11 @@ class DecksController {
 
     static async deleteDeck(req, res) {
         const deckId = req.params.id;
+        const userId = req.user._id;
         try {
-            await Deck.findOneAndDelete({ _id: deckId });
-            res.status(200).send({ success: `Carta apagada` })
+            const deletedDeck = await Deck.findOneAndDelete({ _id: deckId, user: userId });
+            if (deletedDeck) res.status(200).send({ success: `Baralho apagado` });
+            else res.status(403).send({ error: "Não tem permisão para apagar este baralho" })
         } catch (error) {
             res.status(400).send(error)
         }
@@ -50,18 +53,24 @@ class DecksController {
 
     static async updateDeck(req, res) {
         const deckId = req.params.id;
+        const userId = req.user._id;
+        const cards = req.body.cards;
         const update = {
             name: req.body.name,
-            user: req.user._id,
+            user: userId,
             cards: []
         }
         for (let i = 0; i < cards.length; i++) {
-            newDeck.cards.push(cards[i]._id)
+            update.cards.push(cards[i]._id)
         }
 
         try {
-            await Deck.findByIdAndUpdate(deckId, update);
-            res.status(203).send({ success: `Varalho ${req.body.name} atualizado` });
+            const updatedDeck = await Deck.findOneAndUpdate({ _id: deckId, user: userId }, update).populate("cards");
+            if (updatedDeck) {
+                res.status(203).send({ success: `Baralho ${req.body.name} atualizado`, deck: updatedDeck });
+            } else {
+                res.status(403).send({ error: "Não tem permisão para atualizar este baralho" })
+            }
         } catch (error) {
             res.status(400).send(error)
         }
